protect deleting wallet

+++ src/pages/api/wallet/[id].js
@@ -12,6 +12,24 @@ export default async function handler(req, res) {
     case "DELETE":
       // Delete a wallet
       try {
+        const foundWallet = await prisma.wallet.findUnique({
+          where: {
+            id: req.query.id,
+          },
+        });
+
+        if (foundWallet.user_id !== session.user.id) {
+          return res.status(401).json({ message: "Unauthorized" });
+        }
+
+        const transactions = await prisma.transaction.count({
+          where: {
+            wallet_id: req.query.id,
+          },
+        });
+
+        if (transactions > 0) return res.status(400).json({ message: "Cannot delete wallet with transactions" });
+
         const deletedWallet = await prisma.wallet.delete({
           where: {
             id: req.query.id,
